agents/data_analysis/requirements.txt
pika==1.3.2
requests==2.31.0
prometheus-client==0.17.1
python-dotenv==1.0.0
pandas==2.0.2
numpy==1.24.3
scikit-learn==1.2.2